Reorganize aiframe-metrics to support Scala.js
