+++ airframe-metrics/jvm/src/main/scala/wvlet/airframe/metrics/TimeParser.scala
@@ -1,3 +1,16 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package wvlet.airframe.metrics
 
 import java.time._
@@ -51,7 +64,7 @@ object TimeParser extends LogSupport {
     loop(zonedDateTimePatterns.toList)
   }
 
-  def parseAtLocalTimeZone(s: String) = parse(s, TimeWindow.systemZone)
+  def parseAtLocalTimeZone(s: String) = parse(s, systemTimeZone)
 
   def parse(s: String, zone: ZoneOffset): Option[ZonedDateTime] = {
     parseLocalDateTime(s, zone)
@@ -59,4 +72,5 @@ object TimeParser extends LogSupport {
         parseZonedDateTime(s)
       }
   }
+
 }
