+++ airframe-metrics/jvm/src/main/scala/wvlet/airframe/metrics/TimeStampFormatter.scala
@@ -1,3 +1,16 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package wvlet.airframe.metrics
 
 import java.time.format.{DateTimeFormatterBuilder, SignStyle}
@@ -48,13 +61,13 @@ object TimeStampFormatter {
     humanReadableTimestampFormatter.format(time)
   }
 
-  def formatTimestamp(timeMillis: Long, zone: ZoneOffset = TimeWindow.systemZone): String = {
+  def formatTimestamp(timeMillis: Long, zone: ZoneOffset = systemTimeZone): String = {
     val timestamp = ZonedDateTime.ofInstant(Instant.ofEpochMilli(timeMillis), zone)
     humanReadableTimestampFormatter.format(timestamp)
   }
 
   def formatTimestampWithNoSpaace(timeMillis: Long): String = {
-    val timestamp = ZonedDateTime.ofInstant(Instant.ofEpochMilli(timeMillis), TimeWindow.systemZone)
+    val timestamp = ZonedDateTime.ofInstant(Instant.ofEpochMilli(timeMillis), systemTimeZone)
     noSpaceTimestampFormat.format(timestamp)
   }
