+++ airframe-metrics/jvm/src/main/scala/wvlet/airframe/metrics/TimeVector.scala
