+++ airframe-metrics/jvm/src/test/scala/wvlet/airframe/metrics/TimeParserTest.scala
@@ -13,8 +13,7 @@
  */
 package wvlet.airframe.metrics
 
-import java.time.{ZoneOffset, ZonedDateTime}
-import java.util.TimeZone
+import java.time.ZonedDateTime
 
 import wvlet.airframe.AirframeSpec
 
@@ -24,7 +23,7 @@ import wvlet.airframe.AirframeSpec
 class TimeParserTest extends AirframeSpec {
 
   def parse(str: String, expected: String) {
-    val z   = TimeParser.parse(str, TimeWindow.UTC)
+    val z   = TimeParser.parse(str, UTC)
     val ans = ZonedDateTime.parse(expected)
 
     if (z.isEmpty) {
@@ -60,20 +59,6 @@ class TimeParserTest extends AirframeSpec {
       // Datetime without time zone
       parse("2016-12-01 08:00:01", "2016-12-01T08:00:01Z")
       parse("2016-12-01 08:00:01", "2016-12-01T08:00:01Z")
-    }
-
-    "use proper time zone" in {
-      val default = TimeZone.getDefault
-      try {
-        TimeZone.setDefault(TimeZone.getTimeZone("UTC"))
-        val t = TimeParser.parse("2017-04-04", ZoneOffset.of("-07:00"))
-        info(t)
-        val w = TimeWindow.withTimeZone("PDT")
-        val d = w.parse("3d/2017-04-07")
-        info(d)
-      } finally {
-        TimeZone.setDefault(default)
-      }
     }
 
   }
