+++ airframe-metrics/shared/src/test/scala/wvlet/airframe/metrics/DataSizeTest.scala
